EventData: rename neutral parameters header
