+++ Core/src/EventData/CMakeLists.txt
@@ -1,6 +1,6 @@
 target_sources_local(
   ActsCore
   PRIVATE
-    NeutralParameters.cpp
+    NeutralTrackParameters.cpp
     TrackParameters.cpp
 )
