+++ Core/src/Material/SurfaceMaterialMapper.cpp
@@ -7,7 +7,8 @@
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 #include "Acts/Material/SurfaceMaterialMapper.hpp"
-#include "Acts/EventData/NeutralParameters.hpp"
+
+#include "Acts/EventData/NeutralTrackParameters.hpp"
 #include "Acts/Material/BinnedSurfaceMaterial.hpp"
 #include "Acts/Material/ProtoSurfaceMaterial.hpp"
 #include "Acts/Propagator/ActionList.hpp"
