+++ Tests/IntegrationTests/Fatras/FatrasSimulationTests.cpp
@@ -11,7 +11,7 @@
 #include <boost/test/unit_test.hpp>
 #include <random>
 
-#include "Acts/EventData/NeutralParameters.hpp"
+#include "Acts/EventData/NeutralTrackParameters.hpp"
 #include "Acts/EventData/TrackParameters.hpp"
 #include "Acts/MagneticField/ConstantBField.hpp"
 #include "Acts/Propagator/EigenStepper.hpp"
