+++ Tests/IntegrationTests/PropagationTestHelper.hpp
@@ -10,7 +10,7 @@
 
 #include <limits>
 
-#include "Acts/EventData/NeutralParameters.hpp"
+#include "Acts/EventData/NeutralTrackParameters.hpp"
 #include "Acts/Geometry/GeometryContext.hpp"
 #include "Acts/MagneticField/MagneticFieldContext.hpp"
 #include "Acts/Propagator/DebugOutputActor.hpp"
